Add SDS100 support and do some minor quality of life improvements
