+++ server/go.mod
@@ -7,4 +7,6 @@ require (
 	github.com/gobwas/pool v0.2.0 // indirect
 	github.com/gobwas/ws v1.0.3
 	github.com/golang/glog v0.0.0-20160126235308-23def4e6c14b
+	github.com/sirupsen/logrus v1.6.0
+	github.com/tarm/serial v0.0.0-20180830185346-98f6abe2eb07
 )
